requirements.txt
streamlit
openai
faiss-cpu
pypdf
numpy